Remove Allocator task voting
Removes the unused, dead Allocator task oting code.

The Allocator was originally designed to handle the allocation of many
different kinds of resources, and only move tasks to PENDING after all
allocators had been run. However, in time, no allocator other than for
networks has been written, and there is no forseeable future use case
for another allocator.

In addition, the task voting design was bad; it relied on the individual
allocator components to check if the voting for a particular task had
completed and move it to PENDING if so. This would have spread the
responibility for advancing task state to every single allocator, quite
nearly defeating the purpose of task voting altogether.

Therefore, with poorly designed task voting code serving no function
other than to confuse engineers, it has been removed. If we need it in
the future, we can do it better by rewriting it.

Signed-off-by: Drew Erny <drew.erny@docker.com>