Use watch queue instead of cond for node role
Changes (*Node).waitRole to use a watch queue instead of a sync.Cond
Because a sync.Cond isn't "interruptible" naturally, we ended up having
to set up a kind of complicated goroutine dance to make it work. This
removes that goroutine dance and greatly simplifies the waitRole logic
by using an event queue instead.

Signed-off-by: Drew Erny <drew.erny@docker.com>