Remove inappropriate allocator import
Removes inappropriate import of manager/allocator/cnmallocator into
node/node.go. The consequence of this is that in upstream moby/moby, we
can remove that import as well. manager/allocator/cnmallocator is an
implementation detail, and importing it so high in the stack creates an
unacceptably high degree of coupling.

Signed-off-by: Drew Erny <drew.erny@docker.com>