Add API fields for clean removal.
Removing volumes is a tricky proposition, because it is not sufficient
to simply delete the volume in question. The correct removal steps must
be followed to cleanly remove the volume.

First, to remove a volume from a node, the manager must know
affirmatively that it is no longer in use on that node. If a volume is
sitll in use, then it cannot be unpublished on the controller side. To
solve this problem, a repeated string field is added to NodeDescription,
reporting all volumes active on that node.

Second, to remove a volume from Swarm, or to update it, the volume must
not be active and published anywhere. To facilitate this, volume
availability states are added to the VolumeSpec. These states, analogous
to NodeAvailability, control the usage of the volume.

Signed-off-by: Drew Erny <derny@mirantis.com>