Add removal and retrying volume operations
Actually consists of two changes, which have gotten blended together by
mistake.

First, adds code for removing and deleting Volumes.

Second, Adds code to retry failed Volume operations. This relies on a
sort-of-priority queue, which allows us to schedule retries on a timer
and then handle them as the backoff interval elapses for each operation.

This is probably way over-engineered.

Signed-off-by: Drew Erny <derny@mirantis.com>